moved course badge and button to components
