+++ frontend/src/pages/ProgressionChecker/GridView/CourseBadge/CourseButton/index.jsx
@@ -1,0 +1,3 @@
+import CourseButton from "./CourseButton";
+
+export default CourseButton;
