+++ frontend/src/pages/ProgressionChecker/GridView/CourseBadge/index.jsx
@@ -1,0 +1,3 @@
+import CourseBadge from "./CourseBadge";
+
+export default CourseBadge;
